feat(prettier): use Prettier to determine whether a file is formattable
Instead of relying on the user to supply prettier-atom with a whole list of Atom scopes for each
Prettier parser type, we instead rely on the new getFileInfo method from Prettier and let Prettier
tell us whether or not a file is parseable. This will make future parser support automatic as well
as allow Prettier plugins to work. Unfortunately, because we rely on this new feature from Prettier,
it will mean that users will be forced to upgrade their local Prettier versions to be compatible.

BREAKING CHANGE: Because prettier-atom now relies on the new `getFileInfo` method recently added to
Prettier, if you are having prettier-atom use a local version of Prettier instead of the version
that is packaged with the plugin, you will need to manually update your project's local Prettier
version.